+++ NEWS.md
@@ -1,5 +1,8 @@
 # MagmaClustR (development version)
 
+## Minor
+* Fix 'ID' in hyperposterior() and hyperposterior_clust() when not character
+
 # MagmaClustR 1.0.0
 
 * Initial release